refactor: Move Navbar component to its own file and add tooltips to icons
- Moved the Navbar component into its own dedicated file for better separation of concerns and maintainability
- Added basic tooltips to non-navbar icons (dark mode toggle, sign-in/sign-out buttons)
