+++ package.json
@@ -9,6 +9,7 @@
 	"dependencies": {
 		"@radix-ui/react-dialog": "^1.1.1",
 		"@radix-ui/react-slot": "^1.1.0",
+		"@radix-ui/react-tooltip": "^1.1.3",
 		"@the-collab-lab/shopping-list-utils": "^2.2.0",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
@@ -17,6 +18,7 @@
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-hot-toast": "^2.4.1",
+		"react-icons": "^5.3.0",
 		"react-router-dom": "^6.26.0",
 		"tailwind-merge": "^2.5.2",
 		"tailwindcss-animate": "^1.0.7"
